+++ src/com/intellij/idea/plugin/hybris/system/type/model/IndexKey.java
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -36,6 +37,9 @@ import org.jetbrains.annotations.NotNull;
  */
 public interface IndexKey extends DomElement {
 
+    String ATTRIBUTE = "attribute";
+    String LOWER = "lower";
+
     /**
      * Returns the value of the attribute child.
      * <pre>
@@ -46,7 +50,7 @@ public interface IndexKey extends DomElement {
      * @return the value of the attribute child.
      */
     @NotNull
-    @com.intellij.util.xml.Attribute("attribute")
+    @com.intellij.util.xml.Attribute(ATTRIBUTE)
     @Required
     GenericAttributeValue<String> getAttribute();
 
@@ -61,7 +65,7 @@ public interface IndexKey extends DomElement {
      * @return the value of the lower child.
      */
     @NotNull
-    @com.intellij.util.xml.Attribute("lower")
+    @com.intellij.util.xml.Attribute(LOWER)
     FalseAttributeValue getLower();